Relocate Complex user-defined type.
Move it from `test_block_reduce.py` to `helpers.py` such that it can be
used from other tests that need to exercise user-defined type support.
